Semana 09 andamento

+++ src/main/java/semana09/senac/repository/FabricanteRepository.java
@@ -3,5 +3,5 @@ package semana09.senac.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import semana09.senac.model.Fabricante;
 
-public interface FabricanteRepository extends JpaRepository<Fabricante, Long> {
+public interface FabricanteRepository extends JpaRepository<Fabricante, Integer> {
 }
