+++ src/main/java/semana09/senac/repository/ProdutoRepository.java
@@ -3,5 +3,5 @@ package semana09.senac.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import semana09.senac.model.Produto;
 
-public interface ProdutoRepository extends JpaRepository<Produto, Long> {
+public interface ProdutoRepository extends JpaRepository<Produto, Integer> {
 }
